webapp/requirements.txt
flask==3.1.0
flask-cors==5.0.0
bcrypt==4.2.1
python-dotenv==1.0.1
python-docx==1.1.2
pypdf==5.4.0
git+https://github.com/aabtzu/fiat-lux-agents#egg=fiat-lux-agents
